feat: add clear dome with custom class name for each <div> block & refactor canvas logic
