+++ src/App.scss
@@ -2,5 +2,13 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  //   height: 500px;
+  flex-direction: column;
+}
+
+.outputBlock {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  margin-top: 15px;
 }
